Updates existing tests + fixes those in which too tight timing could cause occasionally failing test runs + moves commonly used fakes into a separate module (helpers/fakes.ts)
