+++ test/ConfigCatClientCacheTests.ts
@@ -3,9 +3,9 @@ import { assert } from "chai";
 import { ConfigCatClient } from "../src/ConfigCatClient";
 import { ConfigCatClientCache } from "../src/ConfigCatClient";
 import { AutoPollOptions, ManualPollOptions } from "../src/ConfigCatClientOptions";
-import { FakeConfigCatKernel, FakeConfigFetcher } from "./ConfigCatClientTests";
 import { isWeakRefAvailable, setupPolyfills } from "../src/Polyfills";
 import { allowEventLoop } from "./helpers/utils";
+import { FakeConfigCatKernel, FakeConfigFetcher } from "./helpers/fakes";
 import "./helpers/ConfigCatClientCacheExtensions";
 
 describe("ConfigCatClientCache", () => {
