+++ test/VariationIdMatrixTests.ts
@@ -2,7 +2,7 @@ import { assert } from "chai";
 import "mocha";
 import { User } from "../src/RolloutEvaluator";
 import * as fs from "fs";
-import { FakeConfigFetcherBase, FakeConfigCatKernel } from "./ConfigCatClientTests";
+import { FakeConfigFetcherBase, FakeConfigCatKernel } from "./helpers/fakes";
 import { AutoPollOptions } from "../src/ConfigCatClientOptions";
 import { IConfigCatClient, ConfigCatClient } from "../src/ConfigCatClient";
